[irsdkcpp] 1.0.9 update.  pkgconfig is now correct, usage updated, other cleanup

+++ ports/irsdkcpp/portfile.cmake
@@ -2,7 +2,7 @@ vcpkg_from_github(
     OUT_SOURCE_PATH SOURCE_PATH
     REPO vrkit-platform/irsdk-cpp
     REF "v${VERSION}"
-    SHA512 efdaafc2badcce7ba88fb943882aa301fa3c9959c8de86a99269ade05f188828c485e16f21241d558b108d48e566702e961a1c5b0bda42e523abcc783bdde5c6
+    SHA512 c702479259d77b3eeec85f5fbe177b040c598f2c0b3812139f4c95a1d3c292eff96cc84baa2c620bbfb2c507c61b4425f2cbeb4cb143235b8b25098ab816796a
     HEAD_REF develop
 )
 
@@ -23,8 +23,13 @@ vcpkg_cmake_config_fixup(CONFIG_PATH lib/cmake/${PORT})
 vcpkg_fixup_pkgconfig()
 vcpkg_copy_pdbs()
 
-file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug/include")
-file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/include/include")
+if (EXISTS "${CURRENT_PACKAGES_DIR}/debug/include")
+    file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug/include")
+endif()
+
+if (EXISTS "${CURRENT_PACKAGES_DIR}/include/include")
+    file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/include/include")
+endif()
 
 configure_file("${CMAKE_CURRENT_LIST_DIR}/usage" "${CURRENT_PACKAGES_DIR}/share/${PORT}/usage" COPYONLY)
 vcpkg_install_copyright(FILE_LIST "${SOURCE_PATH}/LICENSE.md")
